+++ .github/actions/scripts/rerun-job.sh
@@ -25,9 +25,9 @@ fi
 
 echo "rerun job $jobID"
 
-$(curl -L \
+curl -L \
   -X POST \
   -H "Accept: application/vnd.github+json" \
   -H "Authorization: Bearer $GITHUB_TOKEN"\
   -H "X-GitHub-Api-Version: 2022-11-28" \
-  https://api.github.com/repos/$OWNER/$REPO/actions/jobs/$jobID/rerun)
+  https://api.github.com/repos/$OWNER/$REPO/actions/jobs/$jobID/rerun
