+++ .github/workflows/rerun-checklist-management-github-action.yml
@@ -29,4 +29,5 @@ jobs:
         PR_NUMBER: ${{ github.event.issue.number }}
         OWNER: ${{ github.event.repository.owner.login }}
         REPO: ${{ github.event.repository.name }}
-      run: cd ./.github/actions/rerun-checklist-management-github-action && go run main.go fefer
+        JOB_TO_RERUN: checklistsManagement
+      run: cd ./.github/actions/rerun-job && go run main.go
